rework camera examples and make them work regardless of orientation
