--- apps/typegpu-docs/src/content/examples/image-processing/chroma-keying-next/index.html
@@ -1,15 +1,0 @@
-<canvas></canvas>
-<video autoplay playsinline></video>
-
-<style>
-  /* Keep the video "painted" so Safari continues decoding frames. */
-  video {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 1px;
-    height: 1px;
-    opacity: 0;
-    pointer-events: none;
-  }
-</style>
