--- apps/typegpu-docs/src/content/examples/image-processing/chroma-keying-next/index.ts
@@ -1,264 +1,0 @@
-import { rgbToYcbcrMatrix } from '@typegpu/color';
-import tgpu from 'typegpu';
-import * as d from 'typegpu/data';
-
-const rareLayout = tgpu.bindGroupLayout({
-  sampling: { sampler: 'filtering' },
-  color: { uniform: d.vec3f },
-  threshold: { uniform: d.f32 },
-  uvTransform: { uniform: d.mat2x2f },
-});
-
-const frequentLayout = tgpu.bindGroupLayout({
-  inputTexture: { externalTexture: {} },
-});
-
-const VertexOutput = d.struct({
-  position: d.builtin.position,
-  uv: d.location(0, d.vec2f),
-});
-
-const shaderCode = tgpu.resolve({
-  template: /* wgsl */ `
-
-@vertex
-fn main_vert(@builtin(vertex_index) idx: u32) -> VertexOutput {
-  const pos = array(
-    vec2( 1.0,  1.0),
-    vec2( 1.0, -1.0),
-    vec2(-1.0, -1.0),
-    vec2( 1.0,  1.0),
-    vec2(-1.0, -1.0),
-    vec2(-1.0,  1.0),
-  );
-
-  const uv = array(
-    vec2(1.0, 0.0),
-    vec2(1.0, 1.0),
-    vec2(0.0, 1.0),
-    vec2(1.0, 0.0),
-    vec2(0.0, 1.0),
-    vec2(0.0, 0.0),
-  );
-
-  var output: VertexOutput;
-  output.position = vec4(pos[idx], 0.0, 1.0);
-  output.uv = uv[idx];
-  return output;
-}
-
-@fragment
-fn main_frag(@location(0) uv: vec2f) -> @location(0) vec4f {
-  let uv2 = uvTransform * (uv - vec2f(0.5)) + vec2f(0.5);
-  var col = textureSampleBaseClampToEdge(inputTexture, sampling, uv2);
-  let ycbcr = col.rgb * rgbToYcbcrMatrix;
-  let colycbcr = color * rgbToYcbcrMatrix;
-
-  let crDiff = abs(ycbcr.g - colycbcr.g);
-  let cbDiff = abs(ycbcr.b - colycbcr.b);
-  let distance = length(vec2f(crDiff, cbDiff));
-
-  if (distance < pow(threshold, 2)) {
-    col = vec4f();
-  }
-
-  return col;
-}
-
-`,
-  externals: {
-    ...rareLayout.bound,
-    ...frequentLayout.bound,
-    VertexOutput,
-    rgbToYcbcrMatrix,
-  },
-});
-
-const canvas = document.querySelector('canvas') as HTMLCanvasElement;
-const video = document.querySelector('video') as HTMLVideoElement;
-
-if (navigator.mediaDevices.getUserMedia) {
-  video.srcObject = await navigator.mediaDevices.getUserMedia({
-    video: {
-      facingMode: 'user',
-      width: { ideal: 1280 },
-      height: { ideal: 720 },
-      frameRate: { ideal: 60 },
-    },
-  });
-} else {
-  throw new Error('getUserMedia not supported');
-}
-
-const root = await tgpu.init();
-const device = root.device;
-
-const context = canvas.getContext('webgpu') as GPUCanvasContext;
-const presentationFormat = navigator.gpu.getPreferredCanvasFormat();
-
-context.configure({
-  device,
-  format: presentationFormat,
-  alphaMode: 'premultiplied',
-});
-
-const thresholdBuffer = root.createBuffer(d.f32, 0.5).$usage('uniform');
-
-const colorBuffer = root
-  .createBuffer(d.vec3f, d.vec3f(0, 1.0, 0))
-  .$usage('uniform');
-
-const uvTransformBuffer = root
-  .createBuffer(d.mat2x2f, d.mat2x2f(1, 0, 0, 1))
-  .$usage('uniform');
-
-const sampler = device.createSampler({
-  magFilter: 'linear',
-  minFilter: 'linear',
-});
-
-const rareBindGroup = root.createBindGroup(rareLayout, {
-  color: colorBuffer,
-  sampling: sampler,
-  threshold: thresholdBuffer,
-  uvTransform: uvTransformBuffer,
-});
-
-const shaderModule = device.createShaderModule({
-  code: shaderCode,
-});
-
-const renderPipeline = device.createRenderPipeline({
-  layout: device.createPipelineLayout({
-    bindGroupLayouts: [root.unwrap(rareLayout), root.unwrap(frequentLayout)],
-  }),
-  vertex: {
-    module: shaderModule,
-  },
-  fragment: {
-    module: shaderModule,
-    targets: [{ format: presentationFormat }],
-  },
-});
-
-function onVideoChange(size: { width: number; height: number }) {
-  const aspectRatio = size.width / size.height;
-  video.style.height = `${video.clientWidth / aspectRatio}px`;
-  if (canvas.parentElement) {
-    canvas.parentElement.style.aspectRatio = `${aspectRatio}`;
-    canvas.parentElement.style.height =
-      `min(100cqh, calc(100cqw/(${aspectRatio})))`;
-  }
-
-  function setUVTransformForIOS() {
-    const isIOS = /iPad|iPhone|iPod/.test(navigator.userAgent);
-    if (!isIOS) {
-      uvTransformBuffer.write(d.mat2x2f(1, 0, 0, 1));
-      return;
-    }
-
-    const angle = screen.orientation.type;
-
-    let m = d.mat2x2f(1, 0, 0, 1);
-    if (angle === 'portrait-primary') {
-      m = d.mat2x2f(0, -1, 1, 0);
-    } else if (angle === 'portrait-secondary') {
-      m = d.mat2x2f(0, 1, -1, 0);
-    } else if (angle === 'landscape-primary') {
-      m = d.mat2x2f(-1, 0, 0, -1);
-    }
-
-    uvTransformBuffer.write(m);
-  }
-
-  setUVTransformForIOS();
-  window.addEventListener('orientationchange', setUVTransformForIOS);
-}
-
-const renderPassDescriptor: GPURenderPassDescriptor = {
-  colorAttachments: [
-    {
-      view: undefined as unknown as GPUTextureView,
-      clearValue: [1, 1, 1, 1],
-      loadOp: 'clear' as const,
-      storeOp: 'store' as const,
-    },
-  ],
-};
-
-let videoFrameCallbackId: number | undefined;
-let lastFrameSize: { width: number; height: number } | undefined = undefined;
-
-function processVideoFrame(
-  _: number,
-  metadata: VideoFrameCallbackMetadata,
-) {
-  if (video.readyState < 2) {
-    videoFrameCallbackId = video.requestVideoFrameCallback(processVideoFrame);
-    return;
-  }
-
-  const frameWidth = metadata.width;
-  const frameHeight = metadata.height;
-
-  if (
-    !lastFrameSize ||
-    lastFrameSize.width !== frameWidth ||
-    lastFrameSize.height !== frameHeight
-  ) {
-    lastFrameSize = { width: frameWidth, height: frameHeight };
-    onVideoChange(lastFrameSize);
-  }
-
-  (
-    renderPassDescriptor.colorAttachments as [GPURenderPassColorAttachment]
-  )[0].view = context.getCurrentTexture().createView();
-
-  const encoder = device.createCommandEncoder();
-
-  const pass = encoder.beginRenderPass(renderPassDescriptor);
-  pass.setPipeline(renderPipeline);
-  pass.setBindGroup(0, root.unwrap(rareBindGroup));
-  pass.setBindGroup(
-    1,
-    root.unwrap(
-      root.createBindGroup(frequentLayout, {
-        inputTexture: device.importExternalTexture({ source: video }),
-      }),
-    ),
-  );
-  pass.draw(6);
-  pass.end();
-
-  device.queue.submit([encoder.finish()]);
-
-  video.requestVideoFrameCallback(processVideoFrame);
-}
-
-video.requestVideoFrameCallback(processVideoFrame);
-
-// #region Example controls & Cleanup
-
-export const controls = {
-  color: {
-    onColorChange: (value: readonly [number, number, number]) => {
-      colorBuffer.write(d.vec3f(...value));
-    },
-    initial: [0, 1, 0] as const,
-  },
-  threshold: {
-    initial: 0.1,
-    min: 0,
-    max: 1,
-    step: 0.01,
-    onSliderChange: (value: number) => thresholdBuffer.write(value),
-  },
-};
-
-export function onCleanup() {
-  if (videoFrameCallbackId !== undefined) {
-    video.cancelVideoFrameCallback(videoFrameCallbackId);
-  }
-}
-
-// #endregion
