--- apps/typegpu-docs/src/content/examples/image-processing/chroma-keying-next/meta.json
@@ -1,5 +1,0 @@
-{
-  "title": "Chroma Keying (next)",
-  "category": "image-processing",
-  "tags": ["camera", "experimental", "filtering", "MediaStreamTrack API"]
-}
