+++ apps/typegpu-docs/src/content/examples/image-processing/chroma-keying/index.html
@@ -2,7 +2,14 @@
 <video autoplay playsinline></video>
 
 <style>
+  /* Keep the video "painted" so Safari continues decoding frames. */
   video {
-    display: none;
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 1px;
+    height: 1px;
+    opacity: 0;
+    pointer-events: none;
   }
 </style>
