+++ apps/typegpu-docs/src/content/examples/image-processing/chroma-keying/meta.json
@@ -1,5 +1,5 @@
 {
   "title": "Chroma Keying",
   "category": "image-processing",
-  "tags": ["camera", "filtering", "MediaStreamTrack API"]
+  "tags": ["camera", "filtering"]
 }
